Move the TranslationLoader to the Translation component.
