+++ src/Symfony/Component/Translation/composer.json
@@ -24,6 +24,7 @@
         "symfony/dependency-injection": "~3.4|~4.0",
         "symfony/intl": "^2.8.18|^3.2.5|~4.0",
         "symfony/yaml": "~3.3|~4.0",
+        "symfony/finder": "~2.8|~3.0|~4.0",
         "psr/log": "~1.0"
     },
     "conflict": {
